Fix bug in EntityManager.
Fixed a subscript out of bounds error that would occur when creating > 50
GameEntities. This error resulted from areas of code that were not updated when
the way components are indexed in EntityManager was updated. Now when we need to
allocate more GameEntities the appropriate number of Component slots are
created.