Separating RenderSystem into multiple systems.
Separated RenderSystem into ClearSystem, RenderSystem, SkyboxSystem, and
SwapSystem. This was done to simplify the RenderSystem::Update function, and to
allow addition/removal of additional rendering systems. For example, the
addition of a UIRenderSystem between SkyboxSystem and SwapSystem. ClearSystem is
responsible for clearing the screen, RenderSystem now only handles rendering all
models, SkyboxSystem is now responsible for rendering the space skybox, and
SwapSystem is responsible for presenting the IDXGISwapChain.